requirements.txt
kaggle
pandas
numpy
scikit-learn
streamlit
google-generativeai
python-dotenv
matplotlib
seaborn
plotly
